+++ tests/test_custom_checkers/test_possibly_undefined_checker.py
@@ -218,7 +218,8 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
 
         self.checker.visit_module(mod)
         with self.assertAddsMessages(
-            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y)
+            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_y)
 
@@ -237,7 +238,8 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
 
         self.checker.visit_module(mod)
         with self.assertAddsMessages(
-            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y)
+            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_y)
 
@@ -258,7 +260,8 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
 
         self.checker.visit_functiondef(func_node)
         with self.assertAddsMessages(
-            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y)
+            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_y)
 
@@ -279,7 +282,8 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
 
         self.checker.visit_functiondef(func_node)
         with self.assertAddsMessages(
-            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y)
+            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_y)
 
@@ -303,7 +307,8 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
 
         self.checker.visit_functiondef(func_node)
         with self.assertAddsMessages(
-            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y)
+            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_y)
 
@@ -321,7 +326,8 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
 
         self.checker.visit_functiondef(func_node)
         with self.assertAddsMessages(
-            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_x)
+            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_x),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_x)
 
@@ -341,6 +347,7 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
         with self.assertAddsMessages(
             pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y),
             pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_x),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_y)
             self.checker.visit_name(name_node_x)
@@ -364,7 +371,8 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
 
         self.checker.visit_functiondef(func_node)
         with self.assertAddsMessages(
-            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y)
+            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_y),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_y)
 
@@ -417,6 +425,7 @@ class TestPossiblyUndefinedChecker(pylint.testutils.CheckerTestCase):
 
         self.checker.visit_functiondef(func_node)
         with self.assertAddsMessages(
-            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_x)
+            pylint.testutils.MessageTest(msg_id="possibly-undefined", node=name_node_x),
+            ignore_position=True,
         ):
             self.checker.visit_name(name_node_x)