+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Enhancements
+
 ## [2.2.0] - 2021-12-09
+- Added new command line argument `-v/--version`. User can print out current PythonTA version using `python -m python_ta -v`.
 
 ### Enhancements