+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 ### Enhancements
+
 - Added new command line argument `-v/--version`. User can print out current PythonTA version using `python -m python_ta -v`.
- 
+
 ## [2.2.0] - 2021-12-09
 
 ### Enhancements